move ecrMetadataService in view-data
