+++ containers/ecr-viewer/src/app/view-data/services/ecrMetadataService.ts
@@ -1,5 +1,14 @@
 import { Bundle, Observation, Organization } from "fhir/r4";
 
+import { evaluatePractitionerRoleReference } from "@/app/services/evaluateFhirDataService";
+import { formatDateTime } from "@/app/services/formatDateService";
+import {
+  formatAddress,
+  formatCodeableConcept,
+  formatContactPoint,
+  formatName,
+} from "@/app/services/formatService";
+import { getReportabilitySummaries } from "@/app/services/reportabilityService";
 import {
   CompleteData,
   noData,
@@ -18,16 +27,6 @@ import {
 } from "@/app/utils/evaluate";
 import fhirPathMappings from "@/app/utils/evaluate/fhir-paths";
 import { DisplayDataProps } from "@/app/view-data/components/DataDisplay";
-
-import { evaluatePractitionerRoleReference } from "./evaluateFhirDataService";
-import { formatDateTime } from "./formatDateService";
-import {
-  formatAddress,
-  formatCodeableConcept,
-  formatContactPoint,
-  formatName,
-} from "./formatService";
-import { getReportabilitySummaries } from "./reportabilityService";
 
 export interface ReportableConditions {
   [condition: string]: {
